Make service more lean
