+++ src/Importer/ImportablePost.php
@@ -9,7 +9,8 @@ class ImportablePost {
 	/**
 	 * Create the object
 	 *
-	 * @param string $importKey     String that uniquely identifies this post for this provider to determine if post has already been imported.
+	 * @param string $importKey     String that uniquely identifies this post for this provider to determine if post has
+	 *                              already been imported.
 	 * @param mixed  $paginationKey If this is the last post, information needed to fetch the next page.
 	 * @param mixed  $postData      Unparsed data from the service.
 	 */
